Chore : Modal 컴포넌트 레이아웃 분리
